feat: display a message in web browser when EnvCheckers halt app
We use ob_start and ob_end_flush to send a snippet of HTML to the client
when an EnvChecker determines that the environment is invalid.  This
helps in the debugging effort as a pointer to the reason an app won't
run.  The HTML is not sent to the output of console commands, it is not
sent when display_errors is on, when a custom error handler is
registered or when the InvalidEnvException thrown by an EnvChecker's is
caught.