fix: consider $_SERVER when checking the environment
Look for environment variables to validate in both $_ENV and $_SERVER
because a value might be in one, but not the other.  This can happen,
for example, when an env var is set in the real environment and Dotenv
doesn't overwite it.  Also check in both aforementioned globals for
undocumented environment variables.